# api: simplify reload benchmark