No need to set up stores and initialise the specs, since we won't
actually use them.

Also reset the timer, to more accurately get the reload time, not the
initialisation time.

benchstat below shows that the benchmark numbers aren't changed
substantially.

name               old time/op    new time/op    delta
ApiInsertReload-4    61.8ms ± 0%    63.0ms ± 0%   ~     (p=1.000 n=1+1)

name               old alloc/op   new alloc/op   delta
ApiInsertReload-4    28.1MB ± 0%    28.1MB ± 0%   ~     (p=1.000 n=1+1)

name               old allocs/op  new allocs/op  delta
ApiInsertReload-4      390k ± 0%      390k ± 0%   ~     (p=1.000 n=1+1)